+++ changelog.md
@@ -17,6 +17,7 @@
   * Fixes attack of slimes dealing damage every tick. Now they deal damage only when jumping and landing, with the range like melee mobs
 * Added Pathfinding feature
   * Mobs are able to pathfind towards the target even if there's no block directly below them
+  * Fixes mobs walking over open trapdoors
 * Only 50% of mobs will get the xray range override
 * Fisher Mobs
   * Fishing range is now configurable